requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
boto3==1.29.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==4.0.1
Django==3.2
django-allauth==0.54.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==2.0.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pluggy==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==1.7.1
pytest==7.4.3
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.0
wcwidth==0.2.9
mysqlclient==2.2.0
django-cors-headers==4.3.0